directories for additional configs and rules
Adds configure/conf.d/ and configure/rules.d/
as a place for 3rd party packages to put their
hooks.

Files in this directory with the extension .make
will be included by the base makefiles.

conf.d is included near the beginning of configure/RULES_BUILD
to allow variables to be modified before any rules
are encountered.

rules.d is included near the end of configure/RULES_BUILD
to allow additional rules to be injected.