libca: don't set thread local storage handle to zero on cleanup
cacExitHandler used to set caClientCallbackThreadId to zero. This
is fine when the IOC is the only user of the dynamically-loaded
libca. However, when an IOC is used with pyDevSup and pyepics,
pyepics ends up depending on the already-loaded libca. At IOC exit,
pyepics tries to cleanup its resources but can't, since its thread-
local storage handle was being zeroed out, which led to a SEGFAULT.

See discussion on https://epics.anl.gov/tech-talk/2019/msg00778.php